requirements/development.txt
# Development dependencies
-r base.txt

# ==============================================================================
# Debug Tools
# ==============================================================================
django-debug-toolbar>=4.4
django-extensions>=3.2

# ==============================================================================
# Code Quality
# ==============================================================================
ruff>=0.7
black>=24.0
isort>=5.13

# ==============================================================================
# Testing (also in test.txt)
# ==============================================================================
pytest>=8.3
pytest-django>=4.9
factory-boy>=3.3
coverage>=7.6
